Update Earnings Regressions Specifications Table Creation.R
leaving off after assessing the regression results for men using the first stage of men's LFP versus the first stage of the women's LFP

The results are different!

If you pretend that women's LFP determines the censoring for men's earnings and for women's earnings, then men and women seem to have basically the same earnings. This does not reflect the fact that women's opportunities outside of the family are worse.... this proxy didn't work. The model predictions don't reflect reality.