Update Earnings Regressions Specifications Table Creation.R
started writing the code to create the BP values.

I made it ~half way then got bored and wanted to play some chess

Need to recover the T_mom_total and T_dad_total transfer values. I deleted the code to make them in this dupllicate file. I'll have to look at the original code to recover it.

Just need to grab the transfer values and add it into the tibble i made, then incorporate it into the equation for BP